flom/requirements.txt
beautifulsoup4
RPi.GPIO
rpimotorlib
schedule

